firebase handled globally
