+++ app.js
@@ -8,14 +8,23 @@ var logger = require('morgan');
 var cookieParser = require('cookie-parser');
 var bodyParser = require('body-parser');
 var session = require("express-session");
+var sess;
+
+firebase = require("firebase");
+
+var config = {
+    apiKey: process.env.FIREBASE_DEVELOPMENT,
+    authDomain: "playwithme-" + process.env.NODE_ENV + ".firebaseapp.com",
+    databaseURL: "https://playwithme-" + process.env.NODE_ENV + ".firebaseio.com/",
+    messagingSenderId: process.env.FIREBASE_MESSAGING_DEVELOPMENT
+};
+firebase.initializeApp(config);
 
 var routes = require('./routes/index');
 var users = require('./routes/users');
 var events = require('./routes/events');
 
 var app = express();
-var session = require('express-session');
-var sess;
 
 // view engine setup
 app.set('views', path.join(__dirname, 'views'));
